requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pymysql==1.1.0
pydantic==2.3.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
jinja2==3.1.2
pytest==7.4.2
httpx==0.25.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
email-validator==2.0.0
cryptography==41.0.3
mysql-connector-python
